requirements.txt
paho-mqtt==1.6.1
azure-storage-blob==12.16.0
python-dotenv==1.0.0